+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+  * Add ``InstructorLite.Adapters.ChatCompletionsCompatible` adapter. `OpenAI`
+    adapter remains but will simply delegate to it under the hood.
   * Update Gemini adapter default model to `gemini-3.5-flash-lite`
 
 ## v1.2.0